renamed client/server to frontend/backend and first draft of complete instructions
