+++ src/frontend/containers/Authenticate.js
@@ -9,7 +9,7 @@ import axios from 'axios/index';
 import {authenticationSuccess, fetchTodosSuccess} from '../actions';
 
 // eslint-disable-next-line no-undef
-const serverUrl = TODOMVC_SERVER_URL;
+const serverUrl = TODOMVC_BACKEND_URL;
 
 const defaultState = {
     password: '',
