+++ src/frontend/containers/Register.js
@@ -9,7 +9,7 @@ import {Form, FormGroup, ControlLabel, FormControl, Button} from 'react-bootstra
 import {registerSuccess} from '../actions';
 
 // eslint-disable-next-line no-undef
-const serverUrl = TODOMVC_SERVER_URL;
+const serverUrl = TODOMVC_BACKEND_URL;
 
 const defaultState = {
     email: '',
