+++ src/frontend/containers/VerifyRegistration.js
@@ -8,7 +8,7 @@ import axios from 'axios/index';
 import {authenticationSuccess, registerSuccess} from '../actions';
 
 // eslint-disable-next-line no-undef
-const serverUrl = TODOMVC_SERVER_URL;
+const serverUrl = TODOMVC_BACKEND_URL;
 
 const defaultState = {
     verificationCode: ''
