+++ src/frontend/reducers/visibilityFilter.js
