+++ utils/backend-helpers/fetch-todos.js
@@ -2,20 +2,20 @@ const {promisify} = require('util');
 
 const prompt = require('prompt');
 
-const index = require('../../src/server/index');
+const index = require('../../src/backend/index');
 
 const promptGetAsync = promisify(prompt.get);
 const doAuthenticateAsync = promisify(index.doAuthenticate);
-const doDeleteTodo = promisify(index.doDeleteTodo);
+const doFetchTodosAsync = promisify(index.doFetchTodos);
 
 const run = async () => {
-    if (!process.env.XilutionApiKey) {
-        throw new Error('XilutionApiKey environment variable must be set.');
+    if (!process.env.XILUTION_SUBSCRIBER_API_KEY) {
+        throw new Error('XILUTION_SUBSCRIBER_API_KEY environment variable must be set.');
     }
 
     prompt.start();
 
-    const input = await promptGetAsync({
+    const credentials = await promptGetAsync({
         properties: {
             username: {
                 required: true
@@ -24,29 +24,19 @@ const run = async () => {
             password: {
                 hidden: true,
                 required: true
-            },
-            // eslint-disable-next-line sort-keys
-            id: {
-                required: true
             }
         }
     });
 
     const authenticationResponse = await doAuthenticateAsync({
-        body: JSON.stringify({
-            password: input.password,
-            username: input.username
-        })
+        body: JSON.stringify(credentials)
     }, {});
 
     const IdToken = JSON.parse(authenticationResponse.body).IdToken;
 
-    return doDeleteTodo({
+    return doFetchTodosAsync({
         headers: {
             Authorization: IdToken
-        },
-        pathParameters: {
-            id: input.id
         }
     }, {});
 };
