+++ webpack.frontend.config.js
@@ -5,7 +5,7 @@ const HtmlWebPackPlugin = require('html-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = (env) => ({
-    entry: ['babel-polyfill', './src/client/index.js'],
+    entry: ['babel-polyfill', './src/frontend/index.js'],
     module: {
         rules: [
             {
@@ -34,7 +34,7 @@ module.exports = (env) => ({
     },
     output: {
         filename: 'bundle.js',
-        path: path.resolve(__dirname, 'dist/client')
+        path: path.resolve(__dirname, 'dist/frontend')
     },
     plugins: [
         new HtmlWebPackPlugin({
@@ -46,7 +46,7 @@ module.exports = (env) => ({
             filename: '[name].css'
         }),
         new webpack.DefinePlugin({
-            TODOMVC_SERVER_URL: JSON.stringify(env.TODOMVC_SERVER_URL)
+            TODOMVC_BACKEND_URL: JSON.stringify(env.TODOMVC_BACKEND_URL)
         })
     ]
 });
